+++ CHANGELOG.md
@@ -11,9 +11,16 @@ the code was deployed.
 
 ## [Unreleased]
 
+### Added
+
+- Added migration script 51 which adds the `status` and `first_device_live_at` column to clients (CU-86dv9uych).
+
 ### Changed
 
-- Updated dashboard homepage to include and sort by Organization (CU-86dva4dxd)
+- Updated brave-alert-lib to v15.0.6 due to client model change (CU-86dv9uych).
+- Updated edit client dashboard page to add new status and first_device_live_at fields (CU-86dv9uych).
+- Updated submitUpdateClientTest.js test cases with new client fields (CU-86dv9uych).
+- Updated dashboard homepage to include and sort by Organization (CU-86dva4dxd).
 
 ## [14.1.0] - 2024-11-14